fix load_shape type

+++ core/models/base/all_commodity_load_shape_ts.sql
@@ -11,7 +11,7 @@ SELECT
     hour_of_day::INTEGER AS hour_of_day,
     hour_of_year::INTEGER AS hour_of_year,
     load_shape::VARCHAR AS load_shape,
-    load_shape_normalized_fraction::NUMERIC AS load_shape_normalized_fraction
+    load_shape_normalized_fraction::FLOAT AS load_shape_normalized_fraction
 FROM (
     SELECT * FROM openbca_reference.commodity_load_shape_ts
     WHERE (commodity, load_shape) NOT IN (
